+++ src/views/shared/Backdrop.tsx
@@ -38,7 +38,6 @@ const Backdrop: React.FC<Props> = ({
     <div
       className="fullscreen"
       style={{
-        backgroundColor: luminosity > 0 ? "white" : "black",
         opacity: show ? 1 : 0,
         transition: "opacity 150ms ease-in-out",
       }}